// src/shared/api/generated/model/transactionItemResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { TransactionItemResponseChannel } from './transactionItemResponseChannel';

export interface TransactionItemResponse {
  ledgerEntryId?: number;
  transactionNumber?: string;
  occurredAt?: string;
  transactionType?: string;
  withdrawalAmount?: number;
  depositAmount?: number;
  transactionContent?: string;
  balanceAfter?: number;
  channel?: TransactionItemResponseChannel;
  reversed?: boolean;
  reversalId?: number;
}